Version 1.0.2
Squashed commits:
Update docs based on the docs team review.
Minor UI improvements.
Add yamato ci.
Android color ints are usually negative. 0 is the new no value.
Check if read/write is enabled directly on the texture instead of the importer.
Refactor iOS enum names to comply with C# conventions. There is no need to duplicate the name of the name in in each field because the full type name must be used in all cases anyway (unlike in c).
Fix notifications not being shown when body field is not set on iOS 10.
Fix crash when querying notification settings on iOS 10.